feat: add local stack and S3MediaStore implementation (#10)

+++ wet_toast_talk_radio/disc_jockey/media_transcoder.py
@@ -4,12 +4,9 @@ from pathlib import Path
 
 import structlog
 from pydub import AudioSegment
-from tqdm import tqdm
 
 from wet_toast_talk_radio.disc_jockey.config import MediaTranscoderConfig
 from wet_toast_talk_radio.media_store import MediaStore
-
-# https://github.com/jiaaro/pydub#getting-ffmpeg-set-up
 
 logger = structlog.get_logger()
 
@@ -50,13 +47,16 @@ class MediaTranscoder:
         while new_raw_shows:
             show = new_raw_shows.pop()
             batch.append(show)
-            if len(batch) > self._cfg.batch_size or len(new_raw_shows) == 0:
-                logger.info("Processing batch...", count=len(batch), batch=batch)
+            if len(batch) >= self._cfg.batch_size or len(new_raw_shows) == 0:
+                logger.info("Processing batch...", batch_len=len(batch), batch=batch)
                 self._download_raw_shows(batch)
                 self._transcode_downloaded_shows()
                 self._upload_tanscoded_shows()
                 self._cleanup_tmp_files()
                 batch = []
+            logger.info(f"{len(new_raw_shows)} shows left to process")
+
+        logger.info("Media transcoder finished!")
 
     def _find_new_raw_shows(self) -> list[str]:
         """Find new raw shows that have not been transcoded yet"""
@@ -78,6 +78,8 @@ class MediaTranscoder:
         self._media_store.download_raw_shows(show_names, self._raw_shows_dir)
 
     def _transcode_downloaded_shows(self):
+        logger.info("Transcoding downloaded shows ...")
+
         def transcode_show(show: Path, out: Path):
             song = AudioSegment.from_mp3(show)
             song.export(out, format="ogg")
@@ -86,22 +88,14 @@ class MediaTranscoder:
             max_workers=self._cfg.max_transcode_workers
         ) as executor:
             futures = []
-            for raw_show in tqdm(
-                self._raw_shows_dir.iterdir(), desc="Transcoding", unit=" file"
-            ):
+            for raw_show in self._raw_shows_dir.iterdir():
                 if raw_show.is_file() and raw_show.suffix == ".wav":
                     out = self._transcoded_shows_dir.joinpath(
                         raw_show.name
                     ).with_suffix(".ogg")
                     futures.append(executor.submit(transcode_show, raw_show, out))
 
-            # Wait for all the tasks to complete
-            for _ in tqdm(
-                concurrent.futures.as_completed(futures),
-                total=len(futures),
-                desc="Progress",
-            ):
-                pass
+            concurrent.futures.wait(futures)
 
     def _upload_tanscoded_shows(self):
         logger.info("Uploading transcoded shows ...")
@@ -112,12 +106,13 @@ class MediaTranscoder:
         self._media_store.upload_transcoded_shows(shows)
 
     def _cleanup_tmp_files(self):
-        for directory in [
-            self._raw_shows_dir,
-            self._transcoded_shows_dir,
-        ]:
-            for path in directory.iterdir():
-                if path.is_file():
-                    path.unlink()
-                elif path.is_dir():
-                    path.rmdir()
+        if self._cfg.clean_tmp_dir:
+            for directory in [
+                self._raw_shows_dir,
+                self._transcoded_shows_dir,
+            ]:
+                for path in directory.iterdir():
+                    if path.is_file():
+                        path.unlink()
+                    elif path.is_dir():
+                        path.rmdir()
